Accept any wildcard certificate
The validator should not make any judgement about the validity of a wildcard
name in the certificate. As long as the the certificate is syntactically
well-formed and the user trusts the trust anchor, the certificate should be
accepted.

The powers that be were consulted on that matter, here are excerpts from the
discussion:

> There are valid wildcard certs out there for *.psl.tld. The CAB Forum
> Baseline Requirements require that the PSL be checked and taken note of
> but - for good reason - do not _prohibit_ issuance to *.psl.tld. There are
> some circumstances, particularly for the PRIVATE part of the list, where
> such a cert is entirely reasonable.

and

> *.amazon is a perfectly valid cert. As is *.appspot.com, even thought
> appspot.com is in the PSL.

(PSL: publicsuffix list)